+++ examples/expr.rs
@@ -119,7 +119,9 @@ fn setup(mut commands: Commands, mut effects: ResMut<Assets<EffectAsset>>) {
                 gradient: size_gradient,
                 screen_space_size: false,
             })
-            .render(OrientAlongVelocityModifier),
+            .render(OrientModifier {
+                mode: OrientMode::AlongVelocity,
+            }),
     );
 
     commands.spawn((