+++ src/asset.rs
@@ -483,7 +483,15 @@ mod tests {
         .render(ParticleTextureModifier::default())
         .render(ColorOverLifetimeModifier::default())
         .render(SizeOverLifetimeModifier::default())
-        .render(BillboardModifier);
+        .render(OrientModifier {
+            mode: OrientMode::ParallelCameraDepthPlane,
+        })
+        .render(OrientModifier {
+            mode: OrientMode::FaceCameraPosition,
+        })
+        .render(OrientModifier {
+            mode: OrientMode::AlongVelocity,
+        });
 
         assert_eq!(effect.capacity, 4096);
 
@@ -511,7 +519,18 @@ mod tests {
         ParticleTextureModifier::default().apply_render(&mut render_context);
         ColorOverLifetimeModifier::default().apply_render(&mut render_context);
         SizeOverLifetimeModifier::default().apply_render(&mut render_context);
-        BillboardModifier.apply_render(&mut render_context);
+        OrientModifier {
+            mode: OrientMode::ParallelCameraDepthPlane,
+        }
+        .apply_render(&mut render_context);
+        OrientModifier {
+            mode: OrientMode::FaceCameraPosition,
+        }
+        .apply_render(&mut render_context);
+        OrientModifier {
+            mode: OrientMode::AlongVelocity,
+        }
+        .apply_render(&mut render_context);
         // assert_eq!(effect.render_layout, render_layout);
     }